+++ src/test/java/lotto/domain/winning/WinningResultTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.winning;
 
 import static org.assertj.core.api.Assertions.assertThat;
